+++ lib/rubocop/cop/layout/empty_line_after_multiline_condition.rb
@@ -62,7 +62,7 @@ module RuboCop
           return if node.ternary?
 
           if node.modifier_form?
-            check_condition(node.condition) unless next_sibling_empty?(node)
+            check_condition(node.condition) if right_sibling(node)
           else
             check_condition(node.condition)
           end
@@ -74,7 +74,7 @@ module RuboCop
         alias on_until on_while
 
         def on_while_post(node)
-          return if next_sibling_empty?(node)
+          return unless right_sibling(node)
 
           check_condition(node.condition)
         end
@@ -116,9 +116,10 @@ module RuboCop
           processed_source[line].blank?
         end
 
-        def next_sibling_empty?(node)
-          next_sibling = node.parent.children[node.sibling_index + 1]
-          next_sibling.nil?
+        def right_sibling(node)
+          return unless node.parent
+
+          node.parent.children[node.sibling_index + 1]
         end
 
         def multiline_when_condition?(when_node)